chore(release): release version 0.4.0
