+++ app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "qri-webapp",
   "productName": "qri",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "qri (\"query\") frontend application",
   "main": "./main.prod.js",
   "author": {
