+++ lib/index.js
@@ -1,12 +1,11 @@
-/* globals __BUILD__, self */
+/* globals __BUILD__ */
 import React from 'react'
 import { render } from 'react-dom'
 import { AppContainer as ReactHotAppContainer } from 'react-hot-loader'
 import Root from './containers/Root'
 import configureStore from './store/configureStore'
 import './app.global.scss'
-
-import * as monaco from 'monaco-editor'
+import './monaco'
 
 const store = configureStore()
 
@@ -49,24 +48,4 @@ if (__BUILD__.MODE === 'production') {
     page () {},
     load () {}
   }
-}
-
-console.log(self)
-
-self.MonacoEnvironment = {
-	getWorkerUrl: function (moduleId, label) {
-		if (label === 'json') {
-			return './json.worker.bundle.js';
-		}
-		if (label === 'css') {
-			return './css.worker.bundle.js';
-		}
-		if (label === 'html') {
-			return './html.worker.bundle.js';
-		}
-		if (label === 'typescript' || label === 'javascript') {
-			return './ts.worker.bundle.js';
-		}
-		return './editor.worker.bundle.js';
-	}
 }
