+++ webpack.config.webapp.dev.js
@@ -8,6 +8,8 @@ import baseConfig from './webpack.config.base'
 import CheckNodeEnv from './internals/scripts/CheckNodeEnv'
 import MonacoWebpackPlugin from 'monaco-editor-webpack-plugin'
 
+import version from './version'
+
 CheckNodeEnv('development')
 
 const port = process.env.PORT || 2505
@@ -68,7 +70,7 @@ export default merge.smart(baseConfig, {
         'API_URL': JSON.stringify('http://localhost:2503'),
         'STATIC_ASSETS_URL': JSON.stringify('http://localhost:2503'),
         'SEGMENT_KEY': JSON.stringify('not_a_key'),
-        'VERSION': JSON.stringify('0.3.0')
+        'VERSION': JSON.stringify(version)
       }
     })
   ],
